+++ CodeConverter/CSharp/ITypeContext.cs
@@ -5,5 +5,6 @@
         AdditionalInitializers Initializers { get; }
         MethodsWithHandles MethodsWithHandles { get; }
         HoistedNodeState HoistedState { get; }
+        bool Any();
     }
 }