+++ Tests/TestData/MultiFileCharacterization/VBToCSResults/ConvertWholeSolution/WindowsAppVb/Folder/FolderForm.Designer.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Diagnostics;
 using System.Drawing;
-using System.Runtime.CompilerServices;
 using System.Windows.Forms;
 using Microsoft.VisualBasic.CompilerServices;
 
@@ -38,566 +37,199 @@ namespace WindowsAppVb
         private void InitializeComponent()
         {
             var resources = new System.ComponentModel.ComponentResourceManager(typeof(FolderForm));
-            _ToolStrip1 = new ToolStrip();
-            _ToolStripButton1 = new ToolStripButton();
-            _ToolStripButton2 = new ToolStripButton();
-            _ToolStripButton3 = new ToolStripButton();
-            _ToolStripButton4 = new ToolStripButton();
-            _ToolStripButton5 = new ToolStripButton();
-            _ToolStripButton6 = new ToolStripButton();
-            _ToolStripButton7 = new ToolStripButton();
-            _ToolStripButton8 = new ToolStripButton();
-            _ToolStripButton9 = new ToolStripButton();
-            _ToolStripButton10 = new ToolStripButton();
-            _ToolStripButton11 = new ToolStripButton();
-            _ToolStripButton12 = new ToolStripButton();
-            _ToolStripButton13 = new ToolStripButton();
-            _ToolStripButton14 = new ToolStripButton();
-            _ToolStripButton15 = new ToolStripButton();
-            _ToolStrip1.SuspendLayout();
+            ToolStrip1 = new ToolStrip();
+            ToolStripButton1 = new ToolStripButton();
+            ToolStripButton2 = new ToolStripButton();
+            ToolStripButton3 = new ToolStripButton();
+            ToolStripButton4 = new ToolStripButton();
+            ToolStripButton5 = new ToolStripButton();
+            ToolStripButton6 = new ToolStripButton();
+            ToolStripButton7 = new ToolStripButton();
+            ToolStripButton8 = new ToolStripButton();
+            ToolStripButton9 = new ToolStripButton();
+            ToolStripButton10 = new ToolStripButton();
+            ToolStripButton11 = new ToolStripButton();
+            ToolStripButton12 = new ToolStripButton();
+            ToolStripButton13 = new ToolStripButton();
+            ToolStripButton14 = new ToolStripButton();
+            ToolStripButton15 = new ToolStripButton();
+            ToolStrip1.SuspendLayout();
             SuspendLayout();
             // 
             // ToolStrip1
             // 
-            _ToolStrip1.Items.AddRange(new ToolStripItem[] { _ToolStripButton1, _ToolStripButton2, _ToolStripButton3, _ToolStripButton4, _ToolStripButton5, _ToolStripButton6, _ToolStripButton7, _ToolStripButton8, _ToolStripButton9, _ToolStripButton10, _ToolStripButton11, _ToolStripButton12, _ToolStripButton13, _ToolStripButton14, _ToolStripButton15 });
-            _ToolStrip1.Location = new Point(0, 0);
-            _ToolStrip1.Name = "_ToolStrip1";
-            _ToolStrip1.Size = new Size(621, 25);
-            _ToolStrip1.TabIndex = 0;
-            _ToolStrip1.Text = "ToolStrip1";
+            ToolStrip1.Items.AddRange(new ToolStripItem[] { ToolStripButton1, ToolStripButton2, ToolStripButton3, ToolStripButton4, ToolStripButton5, ToolStripButton6, ToolStripButton7, ToolStripButton8, ToolStripButton9, ToolStripButton10, ToolStripButton11, ToolStripButton12, ToolStripButton13, ToolStripButton14, ToolStripButton15 });
+            ToolStrip1.Location = new Point(0, 0);
+            ToolStrip1.Name = "ToolStrip1";
+            ToolStrip1.Size = new Size(621, 25);
+            ToolStrip1.TabIndex = 0;
+            ToolStrip1.Text = "ToolStrip1";
             // 
             // ToolStripButton1
             // 
-            _ToolStripButton1.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton1.Image = My.Resources.Resources.test;
-            _ToolStripButton1.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton1.Name = "_ToolStripButton1";
-            _ToolStripButton1.Size = new Size(23, 22);
-            _ToolStripButton1.Text = "ToolStripButton1";
+            ToolStripButton1.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton1.Image = My.Resources.Resources.test;
+            ToolStripButton1.ImageTransparentColor = Color.Magenta;
+            ToolStripButton1.Name = "ToolStripButton1";
+            ToolStripButton1.Size = new Size(23, 22);
+            ToolStripButton1.Text = "ToolStripButton1";
             // 
             // ToolStripButton2
             // 
-            _ToolStripButton2.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton2.Image = My.Resources.Resources.test2;
-            _ToolStripButton2.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton2.Name = "_ToolStripButton2";
-            _ToolStripButton2.Size = new Size(23, 22);
-            _ToolStripButton2.Text = "ToolStripButton2";
+            ToolStripButton2.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton2.Image = My.Resources.Resources.test2;
+            ToolStripButton2.ImageTransparentColor = Color.Magenta;
+            ToolStripButton2.Name = "ToolStripButton2";
+            ToolStripButton2.Size = new Size(23, 22);
+            ToolStripButton2.Text = "ToolStripButton2";
             // 
             // ToolStripButton3
             // 
-            _ToolStripButton3.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton3.Image = My.Resources.Resources.test3;
-            _ToolStripButton3.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton3.Name = "_ToolStripButton3";
-            _ToolStripButton3.Size = new Size(23, 22);
-            _ToolStripButton3.Text = "ToolStripButton3";
+            ToolStripButton3.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton3.Image = My.Resources.Resources.test3;
+            ToolStripButton3.ImageTransparentColor = Color.Magenta;
+            ToolStripButton3.Name = "ToolStripButton3";
+            ToolStripButton3.Size = new Size(23, 22);
+            ToolStripButton3.Text = "ToolStripButton3";
             // 
             // ToolStripButton4
             // 
-            _ToolStripButton4.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton4.Image = My.Resources.Resource1.test;
-            _ToolStripButton4.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton4.Name = "_ToolStripButton4";
-            _ToolStripButton4.Size = new Size(23, 22);
-            _ToolStripButton4.Text = "ToolStripButton4";
+            ToolStripButton4.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton4.Image = My.Resources.Resource1.test;
+            ToolStripButton4.ImageTransparentColor = Color.Magenta;
+            ToolStripButton4.Name = "ToolStripButton4";
+            ToolStripButton4.Size = new Size(23, 22);
+            ToolStripButton4.Text = "ToolStripButton4";
             // 
             // ToolStripButton5
             // 
-            _ToolStripButton5.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton5.Image = My.Resources.Resource1.test2;
-            _ToolStripButton5.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton5.Name = "_ToolStripButton5";
-            _ToolStripButton5.Size = new Size(23, 22);
-            _ToolStripButton5.Text = "ToolStripButton5";
+            ToolStripButton5.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton5.Image = My.Resources.Resource1.test2;
+            ToolStripButton5.ImageTransparentColor = Color.Magenta;
+            ToolStripButton5.Name = "ToolStripButton5";
+            ToolStripButton5.Size = new Size(23, 22);
+            ToolStripButton5.Text = "ToolStripButton5";
             // 
             // ToolStripButton6
             // 
-            _ToolStripButton6.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton6.Image = My.Resources.Resource1.test3;
-            _ToolStripButton6.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton6.Name = "_ToolStripButton6";
-            _ToolStripButton6.Size = new Size(23, 22);
-            _ToolStripButton6.Text = "ToolStripButton6";
+            ToolStripButton6.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton6.Image = My.Resources.Resource1.test3;
+            ToolStripButton6.ImageTransparentColor = Color.Magenta;
+            ToolStripButton6.Name = "ToolStripButton6";
+            ToolStripButton6.Size = new Size(23, 22);
+            ToolStripButton6.Text = "ToolStripButton6";
             // 
             // ToolStripButton7
             // 
-            _ToolStripButton7.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton7.Image = (Image)resources.GetObject("ToolStripButton7.Image");
-            _ToolStripButton7.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton7.Name = "_ToolStripButton7";
-            _ToolStripButton7.Size = new Size(23, 22);
-            _ToolStripButton7.Text = "ToolStripButton7";
+            ToolStripButton7.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton7.Image = (Image)resources.GetObject("ToolStripButton7.Image");
+            ToolStripButton7.ImageTransparentColor = Color.Magenta;
+            ToolStripButton7.Name = "ToolStripButton7";
+            ToolStripButton7.Size = new Size(23, 22);
+            ToolStripButton7.Text = "ToolStripButton7";
             // 
             // ToolStripButton8
             // 
-            _ToolStripButton8.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton8.Image = (Image)resources.GetObject("ToolStripButton8.Image");
-            _ToolStripButton8.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton8.Name = "_ToolStripButton8";
-            _ToolStripButton8.Size = new Size(23, 22);
-            _ToolStripButton8.Text = "ToolStripButton8";
+            ToolStripButton8.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton8.Image = (Image)resources.GetObject("ToolStripButton8.Image");
+            ToolStripButton8.ImageTransparentColor = Color.Magenta;
+            ToolStripButton8.Name = "ToolStripButton8";
+            ToolStripButton8.Size = new Size(23, 22);
+            ToolStripButton8.Text = "ToolStripButton8";
             // 
             // ToolStripButton9
             // 
-            _ToolStripButton9.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton9.Image = (Image)resources.GetObject("ToolStripButton9.Image");
-            _ToolStripButton9.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton9.Name = "_ToolStripButton9";
-            _ToolStripButton9.Size = new Size(23, 22);
-            _ToolStripButton9.Text = "ToolStripButton9";
+            ToolStripButton9.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton9.Image = (Image)resources.GetObject("ToolStripButton9.Image");
+            ToolStripButton9.ImageTransparentColor = Color.Magenta;
+            ToolStripButton9.Name = "ToolStripButton9";
+            ToolStripButton9.Size = new Size(23, 22);
+            ToolStripButton9.Text = "ToolStripButton9";
             // 
             // ToolStripButton10
             // 
-            _ToolStripButton10.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton10.Image = (Image)resources.GetObject("ToolStripButton10.Image");
-            _ToolStripButton10.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton10.Name = "_ToolStripButton10";
-            _ToolStripButton10.Size = new Size(23, 22);
-            _ToolStripButton10.Text = "ToolStripButton10";
+            ToolStripButton10.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton10.Image = (Image)resources.GetObject("ToolStripButton10.Image");
+            ToolStripButton10.ImageTransparentColor = Color.Magenta;
+            ToolStripButton10.Name = "ToolStripButton10";
+            ToolStripButton10.Size = new Size(23, 22);
+            ToolStripButton10.Text = "ToolStripButton10";
             // 
             // ToolStripButton11
             // 
-            _ToolStripButton11.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton11.Image = (Image)resources.GetObject("ToolStripButton11.Image");
-            _ToolStripButton11.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton11.Name = "_ToolStripButton11";
-            _ToolStripButton11.Size = new Size(23, 22);
-            _ToolStripButton11.Text = "ToolStripButton11";
+            ToolStripButton11.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton11.Image = (Image)resources.GetObject("ToolStripButton11.Image");
+            ToolStripButton11.ImageTransparentColor = Color.Magenta;
+            ToolStripButton11.Name = "ToolStripButton11";
+            ToolStripButton11.Size = new Size(23, 22);
+            ToolStripButton11.Text = "ToolStripButton11";
             // 
             // ToolStripButton12
             // 
-            _ToolStripButton12.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton12.Image = (Image)resources.GetObject("ToolStripButton12.Image");
-            _ToolStripButton12.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton12.Name = "_ToolStripButton12";
-            _ToolStripButton12.Size = new Size(23, 22);
-            _ToolStripButton12.Text = "ToolStripButton12";
+            ToolStripButton12.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton12.Image = (Image)resources.GetObject("ToolStripButton12.Image");
+            ToolStripButton12.ImageTransparentColor = Color.Magenta;
+            ToolStripButton12.Name = "ToolStripButton12";
+            ToolStripButton12.Size = new Size(23, 22);
+            ToolStripButton12.Text = "ToolStripButton12";
             // 
             // ToolStripButton13
             // 
-            _ToolStripButton13.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton13.Image = (Image)resources.GetObject("ToolStripButton13.Image");
-            _ToolStripButton13.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton13.Name = "_ToolStripButton13";
-            _ToolStripButton13.Size = new Size(23, 22);
-            _ToolStripButton13.Text = "ToolStripButton13";
+            ToolStripButton13.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton13.Image = (Image)resources.GetObject("ToolStripButton13.Image");
+            ToolStripButton13.ImageTransparentColor = Color.Magenta;
+            ToolStripButton13.Name = "ToolStripButton13";
+            ToolStripButton13.Size = new Size(23, 22);
+            ToolStripButton13.Text = "ToolStripButton13";
             // 
             // ToolStripButton14
             // 
-            _ToolStripButton14.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton14.Image = (Image)resources.GetObject("ToolStripButton14.Image");
-            _ToolStripButton14.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton14.Name = "_ToolStripButton14";
-            _ToolStripButton14.Size = new Size(23, 22);
-            _ToolStripButton14.Text = "ToolStripButton14";
+            ToolStripButton14.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton14.Image = (Image)resources.GetObject("ToolStripButton14.Image");
+            ToolStripButton14.ImageTransparentColor = Color.Magenta;
+            ToolStripButton14.Name = "ToolStripButton14";
+            ToolStripButton14.Size = new Size(23, 22);
+            ToolStripButton14.Text = "ToolStripButton14";
             // 
             // ToolStripButton15
             // 
-            _ToolStripButton15.DisplayStyle = ToolStripItemDisplayStyle.Image;
-            _ToolStripButton15.Image = (Image)resources.GetObject("ToolStripButton15.Image");
-            _ToolStripButton15.ImageTransparentColor = Color.Magenta;
-            _ToolStripButton15.Name = "_ToolStripButton15";
-            _ToolStripButton15.Size = new Size(23, 22);
-            _ToolStripButton15.Text = "ToolStripButton15";
+            ToolStripButton15.DisplayStyle = ToolStripItemDisplayStyle.Image;
+            ToolStripButton15.Image = (Image)resources.GetObject("ToolStripButton15.Image");
+            ToolStripButton15.ImageTransparentColor = Color.Magenta;
+            ToolStripButton15.Name = "ToolStripButton15";
+            ToolStripButton15.Size = new Size(23, 22);
+            ToolStripButton15.Text = "ToolStripButton15";
             // 
             // FolderForm
             // 
             AutoScaleDimensions = new SizeF(6.0F, 13.0F);
             AutoScaleMode = AutoScaleMode.Font;
             ClientSize = new Size(621, 257);
-            Controls.Add(_ToolStrip1);
+            Controls.Add(ToolStrip1);
             Name = "FolderForm";
             Text = "DesignTime images: 123123 then 9 blank, Runtime: 123123123123111 - check tooltips" + "";
-            _ToolStrip1.ResumeLayout(false);
-            _ToolStrip1.PerformLayout();
+            ToolStrip1.ResumeLayout(false);
+            ToolStrip1.PerformLayout();
             Load += new EventHandler(FolderForm_Load);
             ResumeLayout(false);
             PerformLayout();
         }
 
-        private ToolStrip _ToolStrip1;
-
-        internal ToolStrip ToolStrip1
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStrip1;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStrip1 != null)
-                {
-                }
-
-                _ToolStrip1 = value;
-                if (_ToolStrip1 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton1;
-
-        internal ToolStripButton ToolStripButton1
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton1;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton1 != null)
-                {
-                }
-
-                _ToolStripButton1 = value;
-                if (_ToolStripButton1 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton2;
-
-        internal ToolStripButton ToolStripButton2
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton2;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton2 != null)
-                {
-                }
-
-                _ToolStripButton2 = value;
-                if (_ToolStripButton2 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton3;
-
-        internal ToolStripButton ToolStripButton3
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton3;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton3 != null)
-                {
-                }
-
-                _ToolStripButton3 = value;
-                if (_ToolStripButton3 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton4;
-
-        internal ToolStripButton ToolStripButton4
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton4;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton4 != null)
-                {
-                }
-
-                _ToolStripButton4 = value;
-                if (_ToolStripButton4 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton5;
-
-        internal ToolStripButton ToolStripButton5
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton5;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton5 != null)
-                {
-                }
-
-                _ToolStripButton5 = value;
-                if (_ToolStripButton5 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton6;
-
-        internal ToolStripButton ToolStripButton6
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton6;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton6 != null)
-                {
-                }
-
-                _ToolStripButton6 = value;
-                if (_ToolStripButton6 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton7;
-
-        internal ToolStripButton ToolStripButton7
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton7;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton7 != null)
-                {
-                }
-
-                _ToolStripButton7 = value;
-                if (_ToolStripButton7 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton8;
-
-        internal ToolStripButton ToolStripButton8
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton8;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton8 != null)
-                {
-                }
-
-                _ToolStripButton8 = value;
-                if (_ToolStripButton8 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton9;
-
-        internal ToolStripButton ToolStripButton9
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton9;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton9 != null)
-                {
-                }
-
-                _ToolStripButton9 = value;
-                if (_ToolStripButton9 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton10;
-
-        internal ToolStripButton ToolStripButton10
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton10;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton10 != null)
-                {
-                }
-
-                _ToolStripButton10 = value;
-                if (_ToolStripButton10 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton11;
-
-        internal ToolStripButton ToolStripButton11
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton11;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton11 != null)
-                {
-                }
-
-                _ToolStripButton11 = value;
-                if (_ToolStripButton11 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton12;
-
-        internal ToolStripButton ToolStripButton12
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton12;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton12 != null)
-                {
-                }
-
-                _ToolStripButton12 = value;
-                if (_ToolStripButton12 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton13;
-
-        internal ToolStripButton ToolStripButton13
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton13;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton13 != null)
-                {
-                }
-
-                _ToolStripButton13 = value;
-                if (_ToolStripButton13 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton14;
-
-        internal ToolStripButton ToolStripButton14
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton14;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton14 != null)
-                {
-                }
-
-                _ToolStripButton14 = value;
-                if (_ToolStripButton14 != null)
-                {
-                }
-            }
-        }
-
-        private ToolStripButton _ToolStripButton15;
-
-        internal ToolStripButton ToolStripButton15
-        {
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            get
-            {
-                return _ToolStripButton15;
-            }
-
-            [MethodImpl(MethodImplOptions.Synchronized)]
-            set
-            {
-                if (_ToolStripButton15 != null)
-                {
-                }
-
-                _ToolStripButton15 = value;
-                if (_ToolStripButton15 != null)
-                {
-                }
-            }
-        }
+        private ToolStrip ToolStrip1;
+        private ToolStripButton ToolStripButton1;
+        private ToolStripButton ToolStripButton2;
+        private ToolStripButton ToolStripButton3;
+        private ToolStripButton ToolStripButton4;
+        private ToolStripButton ToolStripButton5;
+        private ToolStripButton ToolStripButton6;
+        private ToolStripButton ToolStripButton7;
+        private ToolStripButton ToolStripButton8;
+        private ToolStripButton ToolStripButton9;
+        private ToolStripButton ToolStripButton10;
+        private ToolStripButton ToolStripButton11;
+        private ToolStripButton ToolStripButton12;
+        private ToolStripButton ToolStripButton13;
+        private ToolStripButton ToolStripButton14;
+        private ToolStripButton ToolStripButton15;
     }
 }